Add spireui.com
SpireUI is a web app I have created to easily create a web app for REST APIs. It uses your OpenAPI spec to generate a basic web app which you can then further customize. I think it would be a useful addition to this page.

It doesn't exactly fit your definition of a GUI Editor (it doesn't actually allow you to edit you OpenAPI spec). But that seemed to be the closest category. Let me know if you think there is a more appropriate category. Thanks!